fix: styled fixed
feat: cardTitle component added for the titles in the /CardsContaienr
