+++ src/components/CardsContainer/CardsContainer.tsx
@@ -1,8 +1,8 @@
 import { useLoader } from '../../hooks/useLoader'
 
-import Skills from '../Skills/Skills'
-// import Projects from '../Projects/Projects'
-import Services from '../Services/Services'
+import Projects from './Projects/Projects'
+import Services from './Services/Services'
+import Skills from './Skills/Skills'
 
 import '../../styles/components/Loaders/loader.css'
 
@@ -18,7 +18,7 @@ export default function CardsContainer(): JSX.Element {
       ) : (
         <>
           <Skills />
-          {/* <Projects /> */}
+          <Projects />
           <Services />
         </>
       )}
