+++ src/components/CardsContainer/Services/Services.styles.ts
@@ -1,14 +1,8 @@
 import styled from 'styled-components'
-import { ipad } from '../../styles/screens'
-import { titleStyle } from '../../styles/mixins/titleStyle'
+import { ipad } from '../../../styles/screens'
 
 export const ServicesStyled = styled.div`
   margin-top: 3rem;
-`
-
-export const ServicesTitle = styled.h2`
-  ${titleStyle}
-  margin-bottom: 2rem;
 `
 
 export const ServicesContainer = styled.div`
