+++ src/components/CardsContainer/Services/ServicesHtml/ServicesHtml.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 // Utils
-import { fadeIn } from '../../../utils/animateCss/animateCss'
-import '../../../styles/rotate.css'
+import { fadeIn } from '../../../../utils/animateCss/animateCss'
+import '../../../../styles/rotate.css'
 import {
   LinkReference,
   ServicesHtmlCard,
